+++ engine/2d/IPostEffect.cpp
@@ -234,7 +234,7 @@ void IPostEffect::CreateTexBuff()
 	//‰æ‘œƒCƒ[ƒW
 	std::vector<UINT> img;
 	img.resize(pixelCount);
-	for (uint32_t i = 0; i < pixelCount; i++)
+	for (size_t i = 0; i < pixelCount; i++)
 	{
 		img[i] = 0xff0000ff;
 	}
